requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.21.0
